Upgrade to Astro v5
[Upgrade guide here](https://docs.astro.build/en/guides/upgrade-to/v5/#updating-existing-collections)
- Renamed src/content/config.ts to src/content.config.ts
- Updated collection declarations to be explicit and use loaders
- Renamed all instances of 'slug' to 'id'
- Fixed broken draft page functionality to use 'prod' instead of
  'production'

Test plan:
- Confirm guides, collections and api docs render with no visual
  regressions
