+++ package.json
@@ -11,15 +11,15 @@
     "integration": "vitest --watch=false --config vitest.integration.config.js"
   },
   "devDependencies": {
-    "@astrojs/markdoc": "^0.11.5",
-    "@astrojs/tailwind": "^5.1.1",
-    "@astrojs/vue": "^4.5.2",
+    "@astrojs/markdoc": "0.12.1",
+    "@astrojs/tailwind": "5.1.3",
+    "@astrojs/vue": "5.0.1",
     "@babel/eslint-parser": "^7.14.7",
     "@eslint/js": "^9.10.0",
     "@hbsnow/rehype-sectionize": "^1.0.7",
     "@headlessui/vue": "^1.6.0",
     "@tailwindcss/typography": "^0.5.15",
-    "astro": "^4.16.7",
+    "astro": "5.0.3",
     "astro-expressive-code": "^0.38.3",
     "eslint": "^9.10.0",
     "eslint-plugin-astro": "^1.2.4",
