+++ src/layouts/BaseLayout.astro
@@ -3,7 +3,6 @@ import AppShell from '../components/AppShell.astro';
 import MessagesBubbleDouble from '../components/icons/MessagesBubbleDouble.vue';
 import Footer from '../components/Footer.astro';
 import '../assets/css/tailwind.css';
-import { getCollection } from '../utils/getCollection';
 
 export function getStaticPaths() {
   return [
