+++ src/navigation/Page.js
@@ -18,9 +18,9 @@ class Page {
   }
 
   static fromContent(content) {
-    const { collection, slug, data: frontmatter, headings } = content;
+    const { collection, id, data: frontmatter, headings } = content;
     return new Page({
-      path: `/${collection}/${slug}`,
+      path: `/${collection}/${id}`,
       title: frontmatter.title,
       nav: frontmatter.nav,
       headings,
