+++ src/navigation/apiNavigation.js
@@ -1,4 +1,5 @@
 import { getCollection } from '../utils/getCollection';
+import { render } from 'astro:content';
 import Navigation from '../navigation/Navigation';
 
 const nav = [
@@ -22,7 +23,7 @@ const nav = [
 
 const collections = await getCollection('api');
 const content = await Promise.all(collections.map(async page => {
-  const { headings } = await page.render();
+  const { headings } = await render(page);
   page.headings = headings.filter(heading => heading.depth === 2);
   return page;
 }));
