+++ src/pages/index.astro
@@ -2,7 +2,6 @@
 import BaseLayout from '../layouts/BaseLayout.astro';
 import FeaturedContentCard from '../components/FeaturedContentCard.astro';
 import { getEntry } from 'astro:content';
-import { getCollection } from '../utils/getCollection';
 import navigation from '../navigation/guideNavigation';
 const paymentFlowsGuide = await getEntry('guides', 'payment-flows');
 const farmlandsPortal = await getEntry('guides', 'farmlands-portal');
@@ -40,21 +39,21 @@ const img = '/homepage-preview.png';
         title={paymentFlowsGuide.data.title}
         description={paymentFlowsGuide.data.description}
         img={paymentFlowsGuide.data.img}
-        slug={paymentFlowsGuide.slug}
+        id={paymentFlowsGuide.id}
         type="guide"
       />
       <FeaturedContentCard
         title={farmlandsPortal.data.title}
         description={farmlandsPortal.data.description}
         img={farmlandsPortal.data.img}
-        slug={farmlandsPortal.slug}
+        id={farmlandsPortal.id}
         type="guide"
       />
       <FeaturedContentCard
         title={farmlandsPosIntegrationGuide.data.title}
         description={farmlandsPosIntegrationGuide.data.description}
         img={farmlandsPosIntegrationGuide.data.img}
-        slug={farmlandsPosIntegrationGuide.id}
+        id={farmlandsPosIntegrationGuide.id}
         type="guide"
       />
       <FeaturedContentCard
