+++ src/utils/getCollection.js
@@ -2,6 +2,6 @@ import { getCollection as get } from 'astro:content';
 
 export async function getCollection(collection) {
   return await get(collection, ({ data }) => {
-    return import.meta.env.MODE !== 'production' || !data.draft;
+    return import.meta.env.MODE !== 'prod' || !data.draft;
   });
 }
